Move internal functions to _util.py
This also avoids having to use "#type:ignore" when importing them on tests.
